+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
   for macOS users (starting with HTTPie 0.9.4.).
 * Added the ability to unset a request header with ``Header:``, and send an
   empty value with ``Header;``.
+* Added ``--default-scheme <URL_SCHEME>``.
 
 
 `0.9.4`_ (2016-07-01)